+++ src/main/java/me/drex/fafpatch/impl/block/StateCopyBlock.java
@@ -1,0 +1,26 @@
+package me.drex.fafpatch.impl.block;
+
+import eu.pb4.factorytools.api.block.FactoryBlock;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+import xyz.nucleoid.packettweaker.PacketContext;
+
+public record StateCopyBlock(Block clientBlock) implements FactoryBlock {
+    public static final StateCopyBlock LIGHTNING_ROD = new StateCopyBlock(Blocks.LIGHTNING_ROD);
+
+    @Override
+    public BlockState getPolymerBlockState(BlockState state, PacketContext context) {
+        return clientBlock.withPropertiesOf(state);
+    }
+
+    @Override
+    public boolean isIgnoringBlockInteractionPlaySoundExceptedEntity(BlockState state, ServerPlayer player, InteractionHand hand, ItemStack stack, ServerLevel world, BlockHitResult blockHitResult) {
+        return true;
+    }
+}